requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.2; python_version >= '3.10'
sqlparse==0.5.1; python_version >= '3.8'
celery[redis]==5.4.0; python_version >= '3.8'
psycopg2==2.9.10; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
flower==2.0.1; python_version >= '3.8'
django-cities-light==3.10.1; python_version >= '3.8'
django-phonenumber-field[phonenumberslite]==8.0.0; python_version >= '3.8'
django-bootstrap-datepicker-plus==5.0.5; python_version >= '3.8'
django-select2==8.2.3; python_version >= '3.8'
django-prose-editor==0.9.0; python_version >= '3.10'
django-ipware==7.0.1; python_version >= '3.8'
django-celery-beat==2.7.0; python_version >= '3.8'
django-celery-results==2.5.1; python_version >= '3.8'
django-redis==5.4.0; python_version >= '3.8'